feat: fix Google Sign-In, App Check debug, and optimize location management
- Added Firebase App Check Debug dependency and provider
- Registered SHA fingerprints in Firebase
- Fixed crash and visual borders in HomeFragment
- Implemented intelligent search in ManageLocations
- Optimized locations loading with parallel coroutines
- Standardized XP types to Long across multiple domain services

+++ app/src/main/java/com/futebadosparcas/data/model/User.kt
@@ -63,7 +63,7 @@ data class User(
 
     @get:PropertyName("experience_points")
     @set:PropertyName("experience_points")
-    var experiencePoints: Int = 0,
+    var experiencePoints: Long = 0L,
 
     // Milestones ja conquistados (para evitar duplicacao)
     @get:PropertyName("milestones_achieved")
